refactor: move JSONPatchOperation and related functions to patch.go
